Revert "Add AddressComponent for COVID-19 treatment assessment AB#12434 (#3325)"
This reverts commit 1f415a55247ece38c9c37bf395fddebf9fcf8a33.
