+++ Apps/AdminWebClient/src/ClientApp/package.json
@@ -13,6 +13,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/vue-fontawesome": "^2.0.6",
     "@types/file-saver": "^2.0.5",
+    "bootstrap-vue": "^2.21.2",
     "core-js": "^3.20.2",
     "file-saver": "^2.0.5",
     "inversify": "^5.1.1",
